+++ modules/client-error-reporting/lib/raven-client-factory.js
@@ -49,6 +49,8 @@ function ravenClientFactory(options) {
       // Halley spam, see https://github.com/troupe/gitter-webapp/issues/1056
       'TransportError',
       'BayeuxError',
+      // Elasticsearch, see https://gitlab.com/gitlab-org/gitter/webapp/issues/1950
+      'NoConnections',
       // Random plugins/extensions
       'top.GLOBALS',
       // See: http://blog.errorception.com/2012/03/tale-of-unfindable-js-error. html