+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 19.12.0 - *upcoming*
 
+ - Update `@gitterhq/services@1.21.0` (integrations), https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1078
+    - Adds Discourse integration
+    - Adds The Bug Genie integration
  - `/v1/repo-info?repo=foo%bar` now returns a `204` status code(previously 200) when the given `repo` query parameter can't be found which caused JSON parsing on the frontend to fail, https://gitlab.com/gitlab-org/gitter/webapp/issues/1948
  - Fix `loading-view.js` NPE when hooking iframe `DOMContentLoaded` event, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1208
  - Stop Elasticsearch `NoConnections` error being spammed to Sentry, https://gitlab.com/gitlab-org/gitter/webapp/merge_requests/1209